Refactor/refactor project structure (#33)
* fix: fix gitignore

* refactor: refactor styling

* refactor: refactor components folders

* refactor: delete unused files

* fix: fix styling

* fix: fix paths
