--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-  "cSpell.words": ["reduxjs", "typecheck"]
-}
