+++ src/app/App.tsx
@@ -1,16 +1,16 @@
-import './reset.css';
+import '../shared/ui/reset.css';
 import './App.css';
 import { useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
-import { ThemeProvider } from '@mui/material';
+import { CssBaseline, ThemeProvider } from '@mui/material';
 import { LocalizationProvider } from '@mui/x-date-pickers';
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
-import AppRoutes from '../components/AppRoutes';
+import AppRoutes from '../shared/router/AppRoutes';
 import tokenCache from '../shared/utils/tokenCache';
 import { RootState } from '../shared/store';
 import { useApiClient } from '../api/hooks';
-import theme from '../components/theme';
+import theme from '../shared/ui/theme';
 
 const App = () => {
   const isAuthCustomer = useSelector((state: RootState) => state.auth.isLoggedIn);
@@ -31,6 +31,7 @@ const App = () => {
 
   return (
     <LocalizationProvider dateAdapter={AdapterDayjs} adapterLocale="en">
+      <CssBaseline />
       <ThemeProvider theme={theme}>
         <BrowserRouter>
           <AppRoutes />
