+++ src/components/Registration/types/index.ts
