+++ src/components/login/index.tsx
@@ -13,8 +13,8 @@ import { SubmitHandler, useForm } from 'react-hook-form';
 import { PasswordElement, TextFieldElement } from 'react-hook-form-mui';
 import { MyCustomerSignin } from '@commercetools/platform-sdk';
 import * as yup from 'yup';
-import schemaEmail from '../validation/emailValidation';
-import schemaPass from '../validation/passValidation';
+import schemaEmail from '../../shared/validation/emailValidation';
+import schemaPass from '../../shared/validation/passValidation';
 import Header from '../Header';
 import { useCustomerAuth } from '../../api/hooks';
 import { RoutePaths } from '../../shared/types/enum';
